Code cleanup

+++ Makefile
@@ -127,6 +127,6 @@ anna_yacc.output *.o anna_lex.c anna_lex.h anna_yacc.c anna_yacc.h	\
 anna_float_i.c anna_char_i.c anna_int_i.c anna_string_i.c		\
 anna_complex_i.c anna_object_i.c *.d *.gcov *.gcda *.gcno		\
 anna_vm_short_circut.c
-	rm -r documentation
+	if test -d documentation; then rm -r documentation; fi
 .PHONY: clean
